Adjust to SoundUtils.convertS16ToF32

+++ korau/src/commonMain/kotlin/com/soywiz/korau/sound/SoundUtils.kt
@@ -1,31 +1,31 @@
 package com.soywiz.korau.sound
 
 object SoundUtils {
-	fun convertS16ToF32(channels: Int, input: ShortArray, leftVolume: Int, rightVolume: Int): FloatArray {
+	fun convertS16ToF32(channels: Int, input: ShortArray, leftVolume: Float, rightVolume: Float): FloatArray {
 		val output = FloatArray(input.size * 2 / channels)
-		val optimized = leftVolume == 1 && rightVolume == 1
+		val optimized = leftVolume == 1f && rightVolume == 1f
 		when (channels) {
 			2 ->
 				if (optimized) {
-					for (n in 0 until output.size) output[n] = (input[n] / 32767.0).toFloat()
+					for (n in 0 until output.size) output[n] = (input[n] / 32767f)
 				} else {
 					for (n in 0 until output.size step 2) {
-						output[n + 0] = ((input[n + 0] / 32767.0) * leftVolume).toFloat()
-						output[n + 1] = ((input[n + 1] / 32767.0) * rightVolume).toFloat()
+						output[n + 0] = ((input[n + 0] / 32767f) * leftVolume)
+						output[n + 1] = ((input[n + 1] / 32767f) * rightVolume)
 					}
 				}
 			1 ->
 				if (optimized) {
 					var m = 0
 					for (n in 0 until input.size) {
-						val v = (input[n] / 32767.0).toFloat()
+						val v = (input[n] / 32767f)
 						output[m++] = v
 						output[m++] = v
 					}
 				} else {
 					var m = 0
 					for (n in 0 until input.size) {
-						val sample = (input[n] / 32767.0).toFloat()
+						val sample = (input[n] / 32767f)
 						output[m++] = sample * leftVolume
 						output[m++] = sample * rightVolume
 					}
